backend/requirements-deploy.txt
# Deployment-optimized requirements
# Use this for cloud platforms that have compilation issues

fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart>=0.0.7
python-dotenv==1.0.0
google-generativeai==0.3.2

# Simplified auth dependencies (no native compilation)
pyjwt==2.8.0
python-bcrypt==4.0.1

# Alternative to python-jose if needed
# authlib==1.2.1
